Update welcome message and add conditional view rendering in dashboard
The welcome message in the dashboard has been updated to be specific for the Banting Market dashboard. Additional links and instructions for application have also been included. Moreover, a condition has been added to the dashboard view to only show the welcome component to users with the pre_approved role.