Optimize user application existence check
The code in the ApplicationsController has been updated to check if a user application exists more efficiently. Previously, the entire application object was being pulled into memory, which can be slower and more memory-intensive. Now, using the `exists()` method, we only need to check if there is at least one match, which is faster and consumes less memory.