Update Laravel packages and remove debugging statement
The commit includes updates of Laravel related composer packages to their latest versions, including laravel/framework, laravel/jetstream, and laravel/sanctum. The debugging statement in the apply method of the ApplicationForm has been removed. Additionally, several code adjustments have been made to improve the readability and function of the application, such as the changes in the ApplicationsController and the application-form.blade.php file.