Update relationship between User and Role models
This commit modifies the relationship between the User and Role models, changing it from a HasOne to a BelongsTo relationship. It also adjusts a query in the ApplicationForm class to search for Admin via the newly revised relationship. Additionally, a migration file has been renamed and modified to rename the 'role' column to 'role_id' in the 'users' table.